change create product and tables and migration

+++ app/Http/Requests/SupplierRequest.php
@@ -29,7 +29,7 @@ class SupplierRequest extends FormRequest
     {
         return [
             'name' => ['required', 'string', 'max:255' , 'unique:suppliers,name'],
-            'phone' => ['required', 'integer', 'unique:suppliers,phone'],
+            'phone' => ['required', 'numeric', 'unique:suppliers,phone'],
             
         ];
     }
